refactor(Charts): split chart editor into multiple files
